fix typo req.user.sale --> req.user.salt
Working throught this awesome pluralsight course, I found a small
typo--req.user.sale should be req.user.salt in the updateUser method
of the user controller.  This will add an unwanted property on the
req.user object, but more importantly, will update the password hash
with the wrong salt.  This has the potential to create quite of bit
of frustration to debug.